Renamed component to JournalCard
