+++ components/JournalCard.vue
